% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_code.R
\name{code_to_df}
\alias{code_to_df}
\title{Convert R code to dataframe}
\usage{
code_to_df(file)
}
\arguments{
\item{file}{}
}
\description{
A data frame of all assignments and libraries
}
\examples{
\dontrun{
file <- system.file(package = "codereviewr", "R/test.R") \%>\% 
 code_to_df()
}
}
