% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_code_step.R
\name{identify_code_step}
\alias{identify_code_step}
\title{Title}
\usage{
identify_code_step(code_details, get_line)
}
\arguments{
\item{code_details}{df of code to run through}

\item{get_line}{step number from code_df}
}
\description{
Title
}
