% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_to_df.R
\name{identify_code_steps}
\alias{identify_code_steps}
\title{Identify what is happening in code step}
\usage{
identify_code_steps(code_df, ...)
}
\arguments{
\item{code_df}{parsed data from code_to_df()}
}
\description{
Identify what is happening in code step
}
\examples{
file <- system.file(package = "codereviewr", "R/test.R")
code_df <- code_to_df(file) 
identify_code_steps(code_df)
}
