% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_code_details.R
\name{plot_code_details}
\alias{plot_code_details}
\title{Create chart of dependencies}
\usage{
plot_code_details(code_lineage)
}
\arguments{
\item{...}{file path, passed to code_to_df()}
}
\description{
Create chart of dependencies
}
\examples{
code_lineage <- 
  code_to_df(file) |> 
  identify_code_steps() |>
  track_lineage(
    operation_types = c("assignment", "update"), 
    object_type_exclusion = c("function", "constant"),
    code_exclusion_regex = "<- (switch|system)|^board"
  )
  
plot_code_details(code_lineage)

}
