% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_review.R
\name{plot_output}
\alias{plot_output}
\title{Create chart of dependencies}
\usage{
plot_output(...)
}
\arguments{
\item{...}{file path, passed to code_to_df()}
}
\description{
Create chart of dependencies
}
