% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_code_details.R
\name{plot_code_details}
\alias{plot_code_details}
\title{Create chart of dependencies}
\usage{
plot_code_details(code_steps)
}
\arguments{
\item{...}{file path, passed to code_to_df()}
}
\description{
Create chart of dependencies
}
\examples{
code_steps <- 
  code_to_df(file) |> 
  identify_code_steps(!str_detect(object, "^(board|branch|collapse|env|find|label|node_1|regex|shorten)"))
  
plot_code_details(code_steps)

}
